refactor(runtime): extract resources fetching etc into a REST API client
