+++ .changeset/sunny-monkeys-sing.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+refactor: extract resources fetching etc into a REST API client; removed two undocumented internal exports from the `@makeswift/runtime/client` entrypoint
