+++ packages/runtime/src/api/schema/index.ts
@@ -1,0 +1,1 @@
+export * from './page-pathname-slice'
