+++ packages/runtime/src/api/schema/page-pathname-slice.ts
@@ -1,0 +1,13 @@
+import { z } from 'zod'
+
+export const pagePathnameSlices = z.array(
+  z
+    .object({
+      id: z.string(),
+      basePageId: z.string(),
+      pathname: z.string(),
+      localizedPathname: z.string().optional(),
+      __typename: z.literal('PagePathnameSlice'),
+    })
+    .nullable(),
+)
