+++ packages/runtime/src/client/schema/fonts.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod'
+
+export const fonts = z.object({
+  googleFonts: z.array(
+    z.object({
+      family: z.string(),
+      variants: z.array(z.string()),
+    }),
+  ),
+  siteId: z.string(),
+})
