+++ packages/runtime/src/client/schema/get-page.ts
@@ -1,0 +1,12 @@
+import { z } from 'zod'
+
+export const getPageResult = z.object({
+  pathname: z.string(),
+  locale: z.string(),
+  alternate: z.array(
+    z.object({
+      pathname: z.string(),
+      locale: z.string(),
+    }),
+  ),
+})
