+++ packages/runtime/src/client/schema/index.ts
@@ -1,0 +1,5 @@
+export * from './component-document'
+export * from './fonts'
+export * from './get-page'
+export * from './get-pages'
+export * from './preview-token-payload'
