+++ packages/runtime/src/client/schema/preview-token-payload.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+export const previewTokenPayload = z.object({
+  payload: z.object({
+    version: z.string(),
+  }),
+})
